Separates parser defining syntax from the compiler
